// apps/web/src/components/landing-page/shared/threads-list-mock/types.ts
export interface DemoLabel {
  name: string;
  color: string;
}

export interface DemoMessage {
  authorName: string;
  content: string;
}

export interface DemoThread {
  id: string;
  shortId: number;
  title: string;
  authorName: string;
  assignedUserName?: string;
  priority: number;
  status: number;
  labels: DemoLabel[];
  lastMessage: DemoMessage;
  createdAt: Date;
}